feat: presentation lesson
