--- apps/web/app/modules/Courses/Lesson/LessonItems/File/index.ts
@@ -1,1 +1,0 @@
-export { File } from "./File";
